refactor: improve code quality and maintainability (#210)
* test(e2e): normalize test file extensions to .spec.ts

Playwright convention uses .spec.ts. Rename the 5 files
that were using the Vitest-style .test.ts extension.


* refactor: consolidate ub-moji components into dataset directory

The components/ub-moji/ and components/dataset/ directories both
contained dataset-related components with no meaningful distinction.
Merge them under dataset/ and remove the ub-moji/ directory.

Update all import paths in pages and test files accordingly.


* refactor: remove duplicate utilities and unsafe type assertions

- Export normalizeBasePath from news-routes.ts and remove the
  identical private copy in Header.tsx
- Remove isSupportedLocale from news.ts which duplicated isLocale
  from i18n.ts; use isLocale throughout
- Replace `segments[0] as Locale` and `locales.includes(locale)`
  in Header.tsx with the isLocale type guard


* fix: resolve pre-existing TypeScript errors

- main.astro: guard isLocale call against undefined lang prop
  (lang is optional in Props but isLocale expects string)
- CitationSection.astro: type feedbackTimeout as number | null
  to match the return type of window.setTimeout in browser context


* test: update news-routes tests for renamed buildSectionIndexPath

buildNewsIndexPath was previously renamed to buildSectionIndexPath
but the test file still referenced the old name.


---------
