+++ src/components/dataset/UbMojiDatasetSection.astro
@@ -4,7 +4,7 @@ import { authors } from "@/lib/authors";
 import type { Locale } from "@/lib/i18n";
 import { t } from "@/lib/i18n";
 import { getUbMojiDatasetDetails } from "@/lib/ub-moji-dataset-details";
-import UbMojiDatasetTables from "@/components/ub-moji/UbMojiDatasetTables";
+import UbMojiDatasetTables from "@/components/dataset/UbMojiDatasetTables";
 import CitationSection from "@/components/CitationSection.astro";
 
 type Props = { locale: Locale };
