+++ src/pages/index.astro
@@ -2,7 +2,7 @@
 import Layout from "@/layouts/main.astro";
 import Hero from "@/components/hero/Hero.astro";
 import DatasetOverview from "@/components/dataset/DatasetOverview.astro";
-import UbMojiDatasetSection from "@/components/ub-moji/UbMojiDatasetSection.astro";
+import UbMojiDatasetSection from "@/components/dataset/UbMojiDatasetSection.astro";
 import { defaultLocale } from "@/lib/i18n";
 ---
