+++ tests/e2e/footer-logo.spec.ts
