+++ tests/e2e/hero-section.spec.ts
