+++ tests/e2e/locale-switch.spec.ts
